requirements.txt
# These requirements are only necessary when developing on Oscar.
# You can check for newer versions of these dependencies using this command
#
# $ for x in `yolk -l | awk '{print $1}'`; do yolk -d $x 2>/dev/null; done

# Development
django-debug-toolbar==1.3.0
django-debug-toolbar-template-timings==0.6.4
django-extensions==1.3.11
Werkzeug==0.9.6

# Sandbox
Whoosh==2.6.0
pysolr==3.2.0
requests==2.4.3
uWSGI==2.0.8

# Docs
Sphinx==1.2.3
sphinxcontrib-napoleon==0.2.8
sphinx_rtd_theme==0.1.6

# Testing
coverage==3.7.1
spec==0.11.1
WebTest==2.0.17
django-webtest==1.7.7
pytest==2.7.0
pytest-cache==1.0
pytest-cov==1.8.1
pytest-django==2.8.0
pytest-xdist==1.11
tox==1.8.1
coveralls==0.4.4
behave==1.2.4
django-nose==1.4.1

# Misc
flake8==2.2.5
pyprof2calltree==1.3.2
ipdb==0.8
ipython==2.3.0

# Country data
pycountry==1.8
